crete chart component
