+++ .gitignore
@@ -1,5 +1,5 @@
 /node_modules/
 /MMM-Linky.js
 /node_helper.js
-/data/consumption.json
 /components/
+/data/*.json
